Refactor code, create separate component for trainer list
