+++ package.json
@@ -15,6 +15,7 @@
     "react-router-dom": "^6.7.0",
     "react-scripts": "5.0.1",
     "source-map-loader": "^4.0.1",
+    "styled-components": "^5.3.6",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
